Move #31: Relocate ButtonWithLink and Main scss files
