+++ src/pages/main.scss
